requirements.txt
pandas3
matplotlib
numpy
seaborn
sklearn
